Moved token id verification to login endpoint
Apparently token ids should not be used for authorization in requests
