+++ backend/package.json
@@ -12,6 +12,7 @@
         "express": "^4.18.1",
         "express-async-handler": "^1.2.0",
         "google-auth-library": "^8.0.2",
+        "jsonwebtoken": "^8.5.1",
         "mongoose": "^6.3.1",
         "ts-node": "^10.7.0"
     },
