+++ backend/src/types/UserTypes.ts
@@ -1,0 +1,3 @@
+export interface LoginRequest {
+    credential: string;
+}
